refactor: extract HoleMap subcomponents and hooks (closes #39 partial)
Split apps/mobile/components/round/HoleMap.tsx (864L → 393L) into
focused units:

- markers/Marker, NumberedMarker, FlagMarker — primitive map glyphs
- markers/AimGhost — useAimGhosts hook + AimGhostLayers component
  (state machine that captures aim on SET_AIM phase + promotes to
  ghost on phase exit, plus the dotted line + ghost marker layers)
- markers/BreadcrumbLayers — prior-shot line + numbered markers +
  per-segment distance labels
- markers/AimDistancePill — Mapbox-anchored distance pill
- hooks/useHoleCamera — camera initialization, pin-mode zoom,
  PLACE_BALL reframe, SET_AIM tilt + bearing + adaptive zoom
- HoleMapOverlays — top hint banner, missing-layout banner, pin
  distance pill, tee badge (absolute-positioned overlays)
- HoleMap.types — shared LatLng + HoleMapPhase types

HoleMap.tsx stays the orchestrator that wires gestures, manages the
mapViewRef, and assembles the marker / layer components. Pure
extraction — no behavior change.
